Add files by upload
All cause I wanted to be able to read the temperature and humidity in my walk-in cooler from my work bench. It was just cool to use the PoE feature to eliminate need for a battery and it can be all as passive as can be. No extra RF (if anyone cares as I do).